v1 release

+++ tgbot/views.py
@@ -9,7 +9,7 @@ from tgbot.keyboards import language_inline_button, phone_keyboard, sing_up_appl
     order_markup, change_profile, back_markup, del_order_inline_button, apply_button
 from tgbot.models import B2CUser, B2CCommandText, B2CStep
 from tgbot.utils import command_line, user_update, phone_wrong, wrong_full_name, wrong_data_birthday, create_order, \
-    my_orders, user_profile, create_order_by_conversation
+    my_orders, user_profile, create_order_by_conversation, go_create_order_by_conversation
 
 
 def start(update: Update, context: CallbackContext):
@@ -43,45 +43,50 @@ def main_handler(update: Update, context: CallbackContext):
     step, created = B2CStep.objects.get_or_create(created_by=user_id)
 
     if user:
-        if user.step == 0:
-            # ism kiritish (back)
-            tex = B2CCommandText.objects.filter(text_code=2, lang_code=user.lang).first().text
-            text = command_line(tex)
-            update.message.delete()
-            context.bot.send_message(chat_id=user_id,
-                                     text=text,
-                                     reply_markup=phone_keyboard(user.lang))
+        if user.step.__le__(5):
+            if user.step == 0:
+                # ism kiritish (back)
+                tex = B2CCommandText.objects.filter(text_code=2, lang_code=user.lang).first().text
+                text = command_line(tex)
+                update.message.delete()
+                context.bot.send_message(chat_id=user_id,
+                                         text=text,
+                                         reply_markup=phone_keyboard(user.lang))
 
-        elif user.step == 1:
-            # wrong number or back
-            tex = B2CCommandText.objects.filter(text_code=1, lang_code=user.lang).first().text
-            text = command_line(tex)
-            context.bot.send_message(chat_id=user_id,
-                                     text=text,
-                                     reply_markup=phone_keyboard(user.lang))
-        elif user.step == 2:
-            # Tu'g'ilgan sanangizni kiriting
-            name = update.message.text
-            user = user_update(user_id, 'first_name', name, step=3)
-            tex = B2CCommandText.objects.filter(text_code=6, lang_code=user.lang).first().text
-            text = command_line(tex)
-            update.message.reply_text(text, reply_markup=back_markup(user.lang))
+            elif user.step == 1:
+                # wrong number or back
+                tex = B2CCommandText.objects.filter(text_code=1, lang_code=user.lang).first().text
+                text = command_line(tex)
+                context.bot.send_message(chat_id=user_id,
+                                         text=text,
+                                         reply_markup=phone_keyboard(user.lang))
+            elif user.step == 2:
+                # Tu'g'ilgan sanangizni kiriting
+                name = update.message.text
+                user = user_update(user_id, 'first_name', name, step=3)
+                tex = B2CCommandText.objects.filter(text_code=6, lang_code=user.lang).first().text
+                text = command_line(tex)
+                update.message.reply_text(text, reply_markup=back_markup(user.lang))
 
-        elif user.step == 3:
-            # ro'yhatdan otildi
-            data_birthday = ''
-            temp = update.message.text
-            numbers = [int(x) for x in re.findall(r'[+]?\d+?\d*', temp)]
-            for item in numbers:
-                data_birthday += str(item.__ceil__()) + "-"
-            user = user_update(user_id, 'data_birthday', data_birthday[:-1], step=4)
-            text = user_profile(user_id)
-            update.message.reply_text(text=text, reply_markup=sing_up_apply_markup(user.lang), parse_mode="HTML")
-        elif user.step == 4:
-            text = user_profile(user_id)
-            update.message.reply_text(text=text, reply_markup=sing_up_apply_markup(user.lang), parse_mode="HTML")
-        elif user.step == 5:
-            update.message.reply_text("Оферта", reply_markup=apply_button(user.lang))
+            elif user.step == 3:
+                # ro'yhatdan otildi
+                data_birthday = ''
+                temp = update.message.text
+                numbers = [int(x) for x in re.findall(r'[+]?\d+?\d*', temp)]
+                for item in numbers:
+                    data_birthday += str(item.__ceil__()) + "-"
+                user = user_update(user_id, 'data_birthday', data_birthday[:-1], step=4)
+                text = user_profile(user_id)
+                update.message.reply_text(text=text, reply_markup=sing_up_apply_markup(user.lang), parse_mode="HTML")
+            elif user.step == 4:
+                text = user_profile(user_id)
+                update.message.reply_text(text=text, reply_markup=sing_up_apply_markup(user.lang), parse_mode="HTML")
+            elif user.step == 5:
+                print("yes")
+                document = open(f'media/Оферта.pdf', 'rb')
+                context.bot.send_document(chat_id=user_id, filename='Оферта.pdf', document=document,
+                                          reply_markup=apply_button(user.lang))
+                # update.message.reply_text("Оферта", reply_markup=apply_button(user.lang))
 
         elif user.step == 6:
             msg = update.message.text
@@ -103,6 +108,7 @@ def main_handler(update: Update, context: CallbackContext):
                 orders = my_orders(user.lang, user_id)
                 for order in orders:
                     context.bot.send_message(chat_id=user_id, text=order, parse_mode="HTML",
+                                             disable_web_page_preview=True,
                                              reply_markup=del_order_inline_button(user.lang))
             elif msg in [item.text for item in profile_text] or chat_data:
                 context.chat_data['profile'] = None
@@ -122,7 +128,7 @@ def main_handler(update: Update, context: CallbackContext):
 
         elif user.step == 7:
             msg = update.message.text
-            user = user_update(user_id, 'first_name', msg, step=5)
+            user = user_update(user_id, 'first_name', msg, step=6)
             text = user_profile(user_id)
             context.bot.send_message(chat_id=user_id, text=text, parse_mode="HTML",
                                      reply_markup=change_profile(user.lang))
@@ -131,13 +137,25 @@ def main_handler(update: Update, context: CallbackContext):
             text = command_line(number_tex)
             context.bot.send_message(chat_id=user_id, text=text,
                                      reply_markup=phone_keyboard(user.lang))
+        elif user.step == 9:
+            data_birthday = ''
+            temp = update.message.text
+            numbers = [int(x) for x in re.findall(r'[+]?\d+?\d*', temp)]
+            for item in numbers:
+                data_birthday += str(item.__ceil__()) + "-"
+            user = user_update(user_id, 'data_birthday', data_birthday[:-1], step=6)
+            text = user_profile(user_id)
+            context.bot.send_message(chat_id=user_id, text=text, parse_mode="HTML",
+                                     reply_markup=change_profile(user.lang))
 
 
 def accept(update: Update, context: CallbackContext):
     user_id = update.effective_user.id
     user = B2CUser.objects.get(telegram_id=user_id)
     if user.step.__le__(5):
-        message = update.message.reply_text("Оферта", reply_markup=apply_button(user.lang))
+        document = open(f'media/Оферта.pdf', 'rb')
+        message = context.bot.send_document(chat_id=user_id, filename='Оферта.pdf', document=document,
+                                  reply_markup=apply_button(user.lang))
         B2CUser.objects.filter(telegram_id=user_id).update(step=5, del_message=message.message_id)
     else:
         B2CStep.objects.filter(created_by=user_id).update(step=21)
@@ -174,84 +192,22 @@ def back(update: Update, context: CallbackContext):
         return main_handler(update, context)
     elif user.step == 6:
         step, created = B2CStep.objects.get_or_create(created_by=user_id)
-        if step.step.__le__(12):
+        if step.step.__le__(11):
             step.step -= 1
             step.save()
             create_order(update, context)
-        # if step.step == 1:
-        #     step.step = 0
-        #     step.save()
-        #     start(update, context)
-        # elif step.step == 2:
-        #     step.step = 1
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 3:
-        #     step.step = 2
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 4:
-        #     step.step = 3
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 5:
-        #     step.step = 4
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 6:
-        #     step.step = 5
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 7:
-        #     step.step = 6
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 8:
-        #     step.step = 7
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 9:
-        #     step.step = 8
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 10:
-        #     step.step = 9
-        #     step.save()
-        #     create_order(update, context)
-        # elif step.step == 11:
-        #     step.step = 10
-        #     step.save()
-        #     create_order(update, context)
+        elif step.step == 21:
+            B2CStep.objects.filter(created_by=user_id).update(step=0)
+            update.message.delete()
+            select_action_text = B2CCommandText.objects.filter(text_code=13, lang_code=user.lang).first().text
+            context.bot.send_message(chat_id=user_id, text=select_action_text, reply_markup=order_markup(user.lang))
+            B2CUser.objects.filter(telegram_id=user_id).update(step=6)
         elif step.step == 22:
             step.step = 21
             step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 23:
-            step.step = 21
-            step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 24:
-            step.step = 22
-            step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 25:
-            step.step = 23
-            step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 26:
-            step.step = 24
-            step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 27:
-            step.step = 25
-            step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 28:
-            step.step = 26
-            step.save()
-            create_order_by_conversation(update, context)
-        elif step.step == 29:
-            step.step = 27
+            go_create_order_by_conversation(update, context)
+        elif 22 < step.step < 30:
+            step.step -= 2
             step.save()
             create_order_by_conversation(update, context)
     elif user.step == 8:
